refactor(tauri) moving whitelist warnings to rust, endpoint cleanup
