+++ tauri/src/endpoints/file_system.rs
@@ -10,6 +10,7 @@ use std::io::Write;
 
 use super::cmd::{DirOperationOptions, FileOperationOptions};
 
+#[cfg(read_dir)]
 pub fn read_dir<T: 'static>(
   webview: &mut WebView<'_, T>,
   path: String,
@@ -38,6 +39,7 @@ pub fn read_dir<T: 'static>(
   );
 }
 
+#[cfg(copy_file)]
 pub fn copy_file<T: 'static>(
   webview: &mut WebView<'_, T>,
   source: String,
@@ -65,6 +67,7 @@ pub fn copy_file<T: 'static>(
   );
 }
 
+#[cfg(create_dir)]
 pub fn create_dir<T: 'static>(
   webview: &mut WebView<'_, T>,
   path: String,
@@ -94,6 +97,7 @@ pub fn create_dir<T: 'static>(
   );
 }
 
+#[cfg(remove_dir)]
 pub fn remove_dir<T: 'static>(
   webview: &mut WebView<'_, T>,
   path: String,
@@ -123,6 +127,7 @@ pub fn remove_dir<T: 'static>(
   );
 }
 
+#[cfg(remove_file)]
 pub fn remove_file<T: 'static>(
   webview: &mut WebView<'_, T>,
   path: String,
@@ -143,6 +148,7 @@ pub fn remove_file<T: 'static>(
   );
 }
 
+#[cfg(rename_file)]
 pub fn rename_file<T: 'static>(
   webview: &mut WebView<'_, T>,
   old_path: String,
@@ -170,6 +176,7 @@ pub fn rename_file<T: 'static>(
   );
 }
 
+#[cfg(write_file)]
 pub fn write_file<T: 'static>(
   webview: &mut WebView<'_, T>,
   file: String,
@@ -194,6 +201,7 @@ pub fn write_file<T: 'static>(
   );
 }
 
+#[cfg(read_text_file)]
 pub fn read_text_file<T: 'static>(
   webview: &mut WebView<'_, T>,
   path: String,
@@ -212,6 +220,7 @@ pub fn read_text_file<T: 'static>(
   );
 }
 
+#[cfg(read_binary_file)]
 pub fn read_binary_file<T: 'static>(
   webview: &mut WebView<'_, T>,
   path: String,
